+++ api/reviews.js
@@ -10,6 +10,11 @@ const router = Router()
  * Route to create a new review.
  */
 router.post('/', requireAuth, async function (req, res, next) {
+  const ownerId = parseInt(req.body.userId)
+  if (req.user !== ownerId || !req.user.isAdmin) {
+    res.status(403).send({ error: 'Forbidden' })
+    return
+  }
   try {
     const review = await Review.create(req.body, ReviewClientFields)
     res.status(201).send({ id: review.id })
@@ -39,6 +44,11 @@ router.get('/:reviewId', async function (req, res, next) {
  * Route to update a review.
  */
 router.patch('/:reviewId', requireAuth, async function (req, res, next) {
+  const ownerId = parseInt(req.body.userId)
+  if (req.user !== ownerId || !req.user.isAdmin) {
+    res.status(403).send({ error: 'Forbidden' })
+    return
+  }
   const reviewId = req.params.reviewId
 
   /*
@@ -61,6 +71,11 @@ router.patch('/:reviewId', requireAuth, async function (req, res, next) {
  * Route to delete a review.
  */
 router.delete('/:reviewId', requireAuth, async function (req, res, next) {
+  const ownerId = parseInt(req.body.userId)
+  if (req.user !== ownerId || !req.user.isAdmin) {
+    res.status(403).send({ error: 'Forbidden' })
+    return
+  }
   const reviewId = req.params.reviewId
   const result = await Review.destroy({ where: { id: reviewId }})
   if (result > 0) {
